Skvortsov/feature/show poker combination

+++ server/Code/MorpehFeatures/PlayersFeature/Components/PlayerTurnTimer.cs
@@ -5,5 +5,5 @@ namespace server.Code.MorpehFeatures.PlayersFeature.Components;
 public struct PlayerTurnTimer : IComponent
 {
     public float TimeCurrent;
-    public int TimeMax;
+    public float TimeMax;
 }
